Garments Shop in C++ OOP
Create the classes as shown below, with the indicated functions and attributes for a shopping cart. 
Abstract Base Class Garment with function getname and virtual method calculateprice and a char* name;
Create an Abstract Class Upper  Inherited from Base Class Garment with data member collarsize and member function getcollarsize,

Create an Abstract Class Lower inherited from Garment class with data member waist and member function getwaistsize

Create Class Shirt Inherited from Upper and override calculate price function as per formula listed below. 
Create Pajama and Trouser classes both inherited from Lower and override Function as per formulas listed below.

Requirements: 
⮚ The price is calculated as under, 
• Shirt Price = collar_size*20 + 10; 
• Pajama price = waist*15 + 50; 
• Trouser price = waist*20 + 100; 
• Collar size of shirt = collar_size of upper-1; 
• Waist of pajama = waist – 2; 
• Waist of trouser = waist – 3; 
In main(), provide a menu through which user is allowed to add at most 10 items for his\her shopping cart. The user may also remove any item before the final payment is made. (This may be implanted using an array of size 10). 
The main menu may have the following options, 
• Add an item to shopping cart, 
When this option is selected, the user is shown the list of the available items and is allowed to select one at a time. (at most 10) 
• Remove an item from the cart, 
When this option is selected, the user is shown the list of selected items and is allowed to droop on at a time. 
• Make the payment, 
When this option is selected, the user is shown total bill and the program exits, “thank you for shopping at our store